backend/requirements.txt
fastapi
uvicorn
openai
redis
better_profanity
pydantic
python-dotenv
requests
streamlit
streamlit_extras
google-genai